Refactoring: Relocateing manifest & network config to dedicated files
