+++ app/upgrade_0_11_4.go
@@ -7,13 +7,10 @@ import (
 	"encoding/binary"
 	"encoding/csv"
 	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	wasmTypes "github.com/CosmWasm/wasmd/x/wasm/types"
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	"github.com/cosmos/cosmos-sdk/types"
-	"os"
-	"path"
 	"strings"
 )
 
@@ -24,54 +21,6 @@ var reconciliationData []byte
 var reconciliationDataTestnet []byte
 
 var reconciliationBalancesKey = prefixStringWithLength("balances")
-
-const manifestFilenameBase = "upgrade_manifest.json"
-
-var NetworkInfos = map[string]NetworkConfig{
-	"fetchhub-4": {
-		ReconciliationInfo: &ReconciliationInfo{
-			TargetAddress:   "fetch1tynmzk68pq6kzawqffrqdhquq475gw9ccmlf9gk24mxjjy6ugl3q70aeyd",
-			InputCSVRecords: readInputReconciliationData(reconciliationData),
-		},
-		Contracts: &ContractSet{
-			Reconciliation: &Reconciliation{
-				Addr:     "fetch1tynmzk68pq6kzawqffrqdhquq475gw9ccmlf9gk24mxjjy6ugl3q70aeyd",
-				NewAdmin: getStringPtr("fetch15p3rl5aavw9rtu86tna5lgxfkz67zzr6ed4yhw"),
-				NewLabel: getStringPtr("reconciliation-contract"),
-			},
-			Almanac: &Almanac{
-				ProdAddr: "fetch1mezzhfj7qgveewzwzdk6lz5sae4dunpmmsjr9u7z0tpmdsae8zmquq3y0y",
-			},
-			AName: &AName{
-				ProdAddr: "fetch1479lwv5vy8skute5cycuz727e55spkhxut0valrcm38x9caa2x8q99ef0q",
-			},
-			TokenBridge: &TokenBridge{
-				Addr:     "fetch1qxxlalvsdjd07p07y3rc5fu6ll8k4tmetpha8n",
-				NewAdmin: getStringPtr("fetch15p3rl5aavw9rtu86tna5lgxfkz67zzr6ed4yhw"),
-			},
-		},
-	},
-
-	"dorado-1": {
-		ReconciliationInfo: &ReconciliationInfo{
-			TargetAddress:   "fetch1g5ur2wc5xnlc7sw9wd895lw7mmxz04r5syj3s6ew8md6pvwuweqqavkgt0",
-			InputCSVRecords: readInputReconciliationData(reconciliationDataTestnet),
-		},
-		Contracts: &ContractSet{
-			Reconciliation: &Reconciliation{
-				Addr: "fetch1g5ur2wc5xnlc7sw9wd895lw7mmxz04r5syj3s6ew8md6pvwuweqqavkgt0",
-			},
-			Almanac: &Almanac{
-				ProdAddr: "fetch1tjagw8g8nn4cwuw00cf0m5tl4l6wfw9c0ue507fhx9e3yrsck8zs0l3q4w",
-				DevAddr:  "fetch135h26ys2nwqealykzey532gamw4l4s07aewpwc0cyd8z6m92vyhsplf0vp",
-			},
-			AName: &AName{
-				ProdAddr: "fetch1mxz8kn3l5ksaftx8a9pj9a6prpzk2uhxnqdkwuqvuh37tw80xu6qges77l",
-				DevAddr:  "fetch1kewgfwxwtuxcnppr547wj6sd0e5fkckyp48dazsh89hll59epgpspmh0tn",
-			},
-		},
-	},
-}
 
 func getStringPtr(val string) *string {
 	return &val
@@ -420,157 +369,4 @@ func (app *App) getContractData(ctx types.Context, contractAddr string) (*types.
 	prefixStore := prefix.NewStore(store, contractAddrKey)
 
 	return &addr, &store, &prefixStore, nil
-}
-
-type UpgradeReconciliationTransfer struct {
-	From    string      `json:"from"`
-	EthAddr string      `json:"eth_addr"`
-	Amount  types.Coins `json:"amount"`
-}
-
-type NetworkConfig struct {
-	ReconciliationInfo *ReconciliationInfo
-	Contracts          *ContractSet
-}
-
-type ReconciliationInfo struct {
-	TargetAddress   string
-	InputCSVRecords [][]string
-}
-
-type ContractSet struct {
-	Reconciliation *Reconciliation
-	TokenBridge    *TokenBridge
-	Almanac        *Almanac
-	AName          *AName
-}
-
-type TokenBridge struct {
-	Addr     string
-	NewAdmin *string
-}
-
-type Reconciliation struct {
-	Addr     string
-	NewAdmin *string
-	NewLabel *string
-}
-
-type Almanac struct {
-	DevAddr  string
-	ProdAddr string
-}
-
-type AName struct {
-	DevAddr  string
-	ProdAddr string
-}
-
-type UpgradeManifest struct {
-	Reconciliation *UpgradeReconciliation `json:"reconciliation,omitempty"`
-	Contracts      *Contracts             `json:"contracts,omitempty"`
-}
-
-func initManifest() *UpgradeManifest {
-	return &UpgradeManifest{
-		Reconciliation: &UpgradeReconciliation{
-			Transfers: &UpgradeReconciliationTransfers{
-				Transfers: make([]UpgradeReconciliationTransfer, 0),
-			},
-			ContractState: &UpgradeReconciliationContractState{
-				Balances: make([]UpgradeReconciliationContractStateBalanceRecord, 0),
-			},
-		},
-		Contracts: &Contracts{
-			StateCleaned: make([]string, 0),
-			AdminUpdated: make([]ContractValueUpdate, 0),
-			LabelUpdated: make([]ContractValueUpdate, 0),
-		},
-	}
-}
-
-type Contracts struct {
-	StateCleaned []string              `json:"contracts_state_cleaned,omitempty"`
-	AdminUpdated []ContractValueUpdate `json:"contracts_admin_updated,omitempty"`
-	LabelUpdated []ContractValueUpdate `json:"contracts_label_updated,omitempty"`
-}
-
-type ContractValueUpdate struct {
-	Address string `json:"address"`
-	From    string `json:"from"`
-	To      string `json:"to"`
-}
-
-type ValueUpdate struct {
-	From string `json:"from"`
-	To   string `json:"to"`
-}
-
-type UpgradeReconciliation struct {
-	Transfers     *UpgradeReconciliationTransfers     `json:"transfers,omitempty"`
-	ContractState *UpgradeReconciliationContractState `json:"contract_state,omitempty"`
-}
-
-type UpgradeReconciliationTransfers struct {
-	Transfers                   []UpgradeReconciliationTransfer `json:"transfers"`
-	To                          string                          `json:"to"`
-	AggregatedTransferredAmount types.Coins                     `json:"aggregated_transferred_amount"`
-	NumberOfTransfers           int                             `json:"number_of_transfers"`
-}
-
-type UpgradeReconciliationContractStateBalanceRecord struct {
-	EthAddr  string      `json:"eth_addr"`
-	Balances types.Coins `json:"balances"`
-}
-
-type UpgradeReconciliationContractState struct {
-	Balances                 []UpgradeReconciliationContractStateBalanceRecord `json:"balances"`
-	AggregatedBalancesAmount types.Coins                                       `json:"aggregated_balances_amount"`
-	NumberOfBalanceRecords   int                                               `json:"number_of_balance_records"`
-}
-
-func (app *App) GetManifestFilePath(prefix string) (string, error) {
-	var upgradeFilePath string
-	var err error
-
-	if upgradeFilePath, err = app.UpgradeKeeper.GetUpgradeInfoPath(); err != nil {
-		return "", err
-	}
-
-	upgradeDir := path.Dir(upgradeFilePath)
-
-	manifestFileName := manifestFilenameBase
-	if prefix != "" {
-		manifestFileName = fmt.Sprintf("%s_%s", prefix, manifestFilenameBase)
-	}
-
-	manifestFilePath := path.Join(upgradeDir, manifestFileName)
-
-	return manifestFilePath, nil
-}
-
-func (app *App) SaveManifest(manifest *UpgradeManifest, upgradeLabel string) error {
-	var serialisedManifest []byte
-	var err error
-
-	var manifestFilePath string
-	if manifestFilePath, err = app.GetManifestFilePath(upgradeLabel); err != nil {
-		return err
-	}
-
-	if serialisedManifest, err = json.MarshalIndent(manifest, "", "\t"); err != nil {
-		return fmt.Errorf("failed to marshal manifest: %w", err)
-	}
-
-	var f *os.File
-	if f, err = os.Create(manifestFilePath); err != nil {
-		return fmt.Errorf("failed to create file \"%s\": %w", manifestFilePath, err)
-	}
-	defer f.Close()
-
-	if _, err = f.Write(serialisedManifest); err != nil {
-		return fmt.Errorf("failed to write manifest to the \"%s\" file : %w", manifestFilePath, err)
-	}
-
-	return nil
 }
